# Add bench workflow and benchmarks for encode and decode performance