Performance today:

```
running 6 tests
test decode::benchmark_decode_empty             ... bench:          17 ns/iter (+/- 4)
test decode::benchmark_decode_vector_1234567890 ... bench:         310 ns/iter (+/- 106)
test decode::benchmark_decode_vector_pineapple  ... bench:         284 ns/iter (+/- 46)
test encode::benchmark_encode_empty             ... bench:         144 ns/iter (+/- 31)
test encode::benchmark_encode_vector_1234567890 ... bench:         424 ns/iter (+/- 130)
test encode::benchmark_encode_vector_pineapple  ... bench:         343 ns/iter (+/- 123)
```